requirements.txt
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
django-environ==0.4.5
psycopg2-binary==2.8.6
pyyaml==5.4.1
Pillow==8.1.0
drf-yasg==1.20.0
coverage==5.4